+++ src/tinkoff/course/unit8/service/ModelService.java
@@ -1,7 +1,6 @@
 package unit8.service;
 
 import java.util.List;
-import java.util.Objects;
 import unit8.data.Message;
 import unit8.entity.Brand;
 import unit8.entity.Model;
@@ -21,7 +20,8 @@ public class ModelService {
     }
 
     public Model getModel(String model) {
-        return getModelWithValidation(model);
+        return modelRepository.findByModel(model)
+            .orElseThrow(() -> new ValidationException(Message.MODEL_DOES_NOT_EXIST));
     }
 
     public Model getModelById(long id) {
@@ -30,10 +30,9 @@ public class ModelService {
     }
 
     public void createModel(String model, Brand brand) {
-        modelRepository.findByModel(model)
-            .ifPresent(Model -> {
-                throw new ValidationException(Message.MODEL_ALREADY_EXIST);
-            });
+        if (modelRepository.existsByModel(model)) {
+            throw new ValidationException(Message.MODEL_ALREADY_EXIST);
+        }
 
         Model modelObject = new Model();
         modelObject.setModel(model);
@@ -43,23 +42,19 @@ public class ModelService {
     }
 
     public void updateModel(String oldModel, String newModel) {
-        Model model = getModelWithValidation(oldModel);
-        modelRepository.findByModel(newModel)
-            .ifPresent(Model -> {
-                if (!Objects.equals(Model.getId(), model.getId())) {
-                    throw new ValidationException(Message.MODEL_ALREADY_EXIST);
-                }
-            });
+        Model model = modelRepository.findByModel(oldModel)
+            .orElseThrow(() -> new ValidationException(Message.MODEL_DOES_NOT_EXIST));
+        if (modelRepository.existsByModel(newModel, model.getId())) {
+            throw new ValidationException(Message.MODEL_ALREADY_EXIST);
+        }
+
         modelRepository.updateModel(model.getId(), newModel);
     }
 
     public void deleteModel(String model) {
-        getModelWithValidation(model);
+        if (!modelRepository.existsByModel(model)) {
+            throw new ValidationException(Message.MODEL_DOES_NOT_EXIST);
+        }
         modelRepository.deleteByModel(model);
-    }
-
-    private Model getModelWithValidation(String model) {
-        return modelRepository.findByModel(model)
-            .orElseThrow(() -> new ValidationException(Message.MODEL_DOES_NOT_EXIST));
     }
 }
